make connection more robust
If BREDR is not explicitly disabled in the ADV packet of CLIMBM,
Android silently uses TRANSPORT_AUTO mode, which for some reason could
prefer BRERD over LE, and thus connect fails.
The TRANSPORT parameter is available since API level 21, but it is only
exposed in 23. Before it was public, but hidden (@hide), so it should
be called using the reflection API first.