ClimbService: fix BLE GATT loosing packets on API 19
On API 19 (both Huawei and Sony) in case of multiple writes to
writeCharacteristics, true was returned, but the callback was not called,
violating the API. The same error doesn't seem to be present on API 21-23.

This workaround protects writes with a semaphore. At the moment it is not
checking API level, but goes on all...