ClimbService: reconnect automatically in case of spontaneous disconnect
When reconnect was the responsabilty of the App, it failed when the
App (activity) went to sleep: our disconnect notification got lost and
the App was not reconnecting.

It is not anymore the responsibility of the user to reconnect. The
service is trying to reconnect on its own. Note that disconnect is now
notified only if it is definite, and we are not trying to reconnect
anymore. Since there is no timeout, this might mean that disconnect is
not notified at all.