ClimbLogger main activity lifecycle change
Now unbindService is called from the onPause, so when the activity is in background the bind is removed. The bind is restored in onResume (it was already there).
The mClimbService.deinit(); has been removed form the onDestroy because it was called also on screen rotation. Note: The service is not terminated when the screen is rotated because it is started with startService (see differences between startService and bindService)